refactor: refactor `WinDivertUtils` module
Considering that we may have more filters in the future, not only the name but implementation of the `WinDivertUtils` module are unreasonable:

+ It only supports a fixed layer, limiting its applicability, and its name is ambiguous.

+ It resulted in the main function having a lot of concrete implementations, which made the main function bloated and excessively lengthy.

So I rename the name of this module to `InboundServerNameFilterService` and move implementations into the module itself. It will be easier to understand the purpose of this module and reduce the risk of code conflicts.