fix & refactor: duplicate detection doesn't work
Cause:

`Equal` method or `==` operator doesn't compare the two patterns but `Regex` objects.

Fix:

`Regex.ToString` method will return the pattern. Comparing the patterns instead of the objects solves this problem.

Also, to avoid unnecessary conversion, do duplicate detection on raw pattern. The method that accepts a `Regex` object as an argument is useless, so it's removed.