fix: can't exit strict mode
Cause:

In the past I just release the pointer to object. I thought the GC would destroy the object, but I was wrong. As the task in `OutboundRequestFilterService` wouldn't finished and exited (it had a `while (true)` loop), the object would never be destroyed indeed.

Fix:

Using `CancellationToken` to stop the task solves this problem.