fix: service may not response to challenge
Cause:

There was a mistake in the previous code. To determine if the incoming data starts with the specific prefix indicating that this packet contains challenge number, I used the method `SequenceEqual` which compares all of the incoming data with the prefix. This approach is incorrect and will always return false, meaning that the sniffing service will never response to the challenge actually.

Fix:

Use method `Take` to compare the first five members of incoming data with the prefix. This will ensure that the comparison is accurate and should produce the desired result.